LLM/requirements.txt
pillow>=9.0.0
matplotlib>=3.5.0
pyautogui>=0.9.54
pyperclip>=1.8.2
volcengine-python-sdk[ark]>=1.0.0 